requirements.txt
github3.py==1.0.0a4
python-gitlab==0.13
GitPython==2.0.8
